Define a FailureDiffs folder
This took an unexpected turn. There's a nice interface defined for the
reference folder:
https://github.com/facebook/ios-snapshot-test-case/blob/master/FBSnapshotTestCase/FBSnapshotTestController.h#L72

That's not the case for the failure diffs folder. I had to get creative
so I used the setenv in order to give this failure diffs folder mapping
by default. See here to know how it's used in the FBSnapshotTestCase
project:
https://github.com/facebook/ios-snapshot-test-case/blob/master/FBSnapshotTestCase/FBSnapshotTestController.m#L266-L268